warning solved: signed/unsigned mismatch
git-svn-id: svn://svn.pgadmin.org/trunk@1735 a7884b65-44f6-0310-8a51-81a127f17b15

+++ src/utils/misc.cpp
@@ -374,7 +374,7 @@ wxString FileRead(const wxString &filename, wxWindow *errParent, int format)
                 break;
         }
 
-        int nLen = conv.MB2WC(0, buf, 0);
+        size_t nLen = conv.MB2WC(0, buf, 0);
         if (nLen == (size_t) -1)
         {
             // Format error
